eventstat: new utility to count events with unbounded cardinality
There are cases when we would like to count events with unbounded
cardinality:

 * number of requests per user agents
 * top users/projects who uses the majority of our resources

It's not safe to use monkit for these cases. For example: we can have
unlimited number of different user agents which would fill our memory
with different monkit counters.

This patch suggests an alternative: all the counters are reset after
publishing the current snapshot. But it works only if there is only one
publisher: (if all the counters are reset after publishing, the second
published would see partial result).

For this reason it's more safe to use a dedicated api instead of monkit,
but we can publish the results in the same format, using UDP packets.

udp.go is imported form https://github.com/zeebo/admission to make it
reusable between eventstat and telemetry.

Change-Id: I0bfc21ae33550eaba7d73f9c234595f41a00be5e