proto: extend piece upload with hash algorithm
This patch extends the existing protocol with an optional hash algorithm
field. In the future we will use it to introduce new piece hash
algorithm (requires suppoort from all the storagenodes).

As soon as the existing (default) sha256 is used it's fully backward
compatible.

The main goal is improving the performance with ~30-35%:

https://github.com/storj/storj/issues/4721

This patch adds support for
 * xxh3 (very fast, but not crypto-safe)
 * blake3 (fast, crypto-safe)

The patch introduces two new dependencies:

 * github.com/zeebo/xxh3: 2 Cause BSD
 * github.com/klauspost/cpuid/v2: MIT
 * github.com/zeebo/blake3: CC0 1.0
 * github.com/zeebo/pcg: CC0 1.0
 * github.com/zeebo/assert: CC0 1.0

Change-Id: I0afac3c1f263affdf2345976edc1739e2700cad7